Replace Digest with flatbuf generated one
These two types look exactly the same, expect that the generated one
makes the internal bytes public. This is not particulary problematic,
and has the benefit of not needing to cast or convert any deserialized
digests in the future.
