+++ Cargo.toml
@@ -45,6 +45,7 @@ colored = "2.0.0"
 config = "0.13.1"
 console = "0.15"
 dashmap = "5.4.0"
+data-encoding = "2.3"
 dirs = "5.0"
 dunce = "1.0.4"
 dyn-clone = "1.0"
