+++ crates/spfs-encoding/Cargo.toml
@@ -8,9 +8,9 @@ version = { workspace = true }
 workspace = true
 
 [dependencies]
-data-encoding = "2.3"
+data-encoding = { workspace = true }
 ring = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
+spfs-proto = { path = "../spfs-proto" }
 tokio = { version = "1.20", features = ["io-util", "io-std"] }
 thiserror = { workspace = true }
 miette = { workspace = true }
