+++ crates/spfs-proto/Cargo.toml
@@ -7,11 +7,20 @@ version = { workspace = true }
 [lints]
 workspace = true
 
+[features]
+serde = ["dep:serde"]
+
 [dependencies]
+data-encoding = { workspace = true }
 flatbuffers = { workspace = true }
+futures = { workspace = true }
+miette = { workspace = true }
+serde = { workspace = true, optional = true }
+thiserror = { workspace = true }
 
 [build-dependencies]
 flatc-rust = "0.2"
 
 [dev-dependencies]
+ring = { workspace = true }
 rstest = { version = "0.15.0", default_features = false }
