+++ crates/spfs-proto/build.rs
@@ -1,3 +1,7 @@
+// Copyright (c) Sony Pictures Imageworks, et al.
+// SPDX-License-Identifier: Apache-2.0
+// https://github.com/imageworks/spk
+
 use std::path::Path;
 
 fn main() {
