+++ crates/spfs/src/tracking/manifest.rs
@@ -369,7 +369,7 @@ pub trait BlobHasher {
 #[tonic::async_trait]
 impl BlobHasher for () {
     async fn hash_blob(&self, reader: Pin<Box<dyn BlobRead>>) -> Result<encoding::Digest> {
-        Ok(encoding::Digest::from_async_reader(reader).await?)
+        Ok(encoding::Hasher::hash_async_reader(reader).await?)
     }
 }
